+++ app/src/main/java/com/animan/wholesalemanager/data/local/ProductReport.kt
@@ -3,6 +3,6 @@ package com.animan.wholesalemanager.data.local
 data class ProductReport(
     val productId: String,
     val name: String,
-    val totalQty: Int,
+    val totalQty: Double,
     val totalRevenue: Double
 )
